Fix silent test failure in image cache tests (#56492)
